Restructure Chat V1/V2. Run lint fix
