+++ .eslintrc.json
@@ -14,7 +14,8 @@
     "rules": {
         "object-curly-spacing": ["error", "always"],
         "indent": ["error", 2],
-        "semi": ["error", "always"]
+        "semi": ["error", "always"],
+        "no-unused-vars": "warn"
     },
     "ignorePatterns": ["node_modules/**", "dist/**"]
 }
