+++ sdk/chat/v2/app-chat.js
@@ -1,11 +1,11 @@
 const styles = require('./app-chat.css');
-const {html, render} = require('lit-html');
-const {createRef, ref} = require("lit-html/directives/ref.js");
-const {when} = require("lit-html/directives/when.js");
+const { html, render } = require('lit-html');
+const { createRef, ref } = require("lit-html/directives/ref.js");
+const { when } = require("lit-html/directives/when.js");
 
 class AppChat extends HTMLElement {
   messages = []; // {"type": "received/sent", "message": "string"}
-  #showRoot
+  #showRoot;
   #inputRef = createRef();
   onSendCb = () => undefined;
   me = '';
@@ -13,12 +13,12 @@ class AppChat extends HTMLElement {
 
   constructor() {
     super();
-    this.#showRoot = this.attachShadow({mode: 'open'});
+    this.#showRoot = this.attachShadow({ mode: 'open' });
   }
 
   getTemplate() {
     const renderMessages = () => html`${this.messages.map(message => html`
-        <div class="message ${message.type}-message">${message.message}</div>`)}`
+        <div class="message ${message.type}-message">${message.message}</div>`)}`;
     return html`
         <style>${styles.default.toString()}</style>
         <div class="chat-box">
@@ -33,7 +33,7 @@ class AppChat extends HTMLElement {
                 <button class="send-button" @click="${this.sendMessage.bind(this)}">Send</button>
             </form>
         </div>
-    `
+    `;
   }
 
   sendMessage(event) {
@@ -41,13 +41,13 @@ class AppChat extends HTMLElement {
     const message = this.#inputRef.value.value;
     this.#inputRef.value.value = '';
     this.#inputRef.value.focus();
-    this.messages.push({type: 'sent', message});
-    this.onSendCb({from: this.me, to: this.them, message});
+    this.messages.push({ type: 'sent', message });
+    this.onSendCb({ from: this.me, to: this.them, message });
     this.render();
   }
 
   receiveMessage(message) {
-    this.messages.push({type: 'received', message});
+    this.messages.push({ type: 'received', message });
     this.render();
   }
