+++ sdk/chat/v2/chat_handlers_v2.js
@@ -8,4 +8,4 @@ function renderAppChats(chatsDivId, currentUserEmail) {
   chatsApp.render();
 }
 
-module.exports = {renderAppChats};
+module.exports = { renderAppChats };
